requirements.txt
googletrans==2.4.0
PyQt5==5.9.2
python-docx==0.8.10
odfpy==1.4.0
tika==1.19